+++ common/RELEASE_NOTES.md
@@ -1,5 +1,12 @@
 # dxCommon
 
+## ## 0.5.0 (2021-06-23)
+
+* Adds optional `localizationDir` parameter to `LocalizationDisambiguator` methods for specifying the localization directory that must be used
+* Fixes `prettyFormat` function to handle case clases with public fields in the second parameter list
+* Adds `recursive` parameter (default=`false`) to `FileSource.listing`
+* Caches parent `LocalFileSource` when resolving children
+
 ## 0.4.1 (2021-06-08)
 
 * Fixes issues with local and http path relativization